+++ micro_tracker/__init__.py
@@ -1,4 +1,4 @@
 """Micro Tracker application package."""
 
-__version__ = "2.9.1"
-__release_date__ = "2026-03-24"
+__version__ = "2.9.2"
+__release_date__ = "2026-05-26"